refactor: Move json safe convert functions to a new file - CU-86c0xzmxm
