+++ lib/src/utils/safe_convert.dart
@@ -1,4 +1,8 @@
-/// Convert any dynamic value to int or null safely.
+// Convert any dynamic value to int or null safely.
+import 'dart:developer';
+
+import 'package:flutter/foundation.dart';
+
 int? toIntOrNull(value) {
   if (value == null) return null;
   if (value is int) return value;
@@ -24,11 +28,25 @@ double? toDoubleOrNull(value) {
 bool? toBoolOrNull(value) {
   if (value == null) return null;
   if (value is bool) return value;
-  if (value is int) return value == 0 ? false : true;
-  if (value is double) return value == 0 ? false : true;
+  if (value is int) {
+    return value == 0
+        ? false
+        : value == 1
+            ? true
+            : null;
+  }
+  if (value is double) {
+    return value == 0
+        ? false
+        : value == 1
+            ? true
+            : null;
+  }
   if (value is String) {
-    if (value == "1" || value.toLowerCase() == "true") return true;
-    if (value == "0" || value.toLowerCase() == "false") return false;
+    if (value.toLowerCase() == "true") return true;
+    if (value.toLowerCase() == "false") return false;
+    if (toDoubleOrNull(value) == 1) return true;
+    if (toDoubleOrNull(value) == 0) return false;
   }
   return null;
 }
@@ -46,7 +64,16 @@ String? toStringOrNull(value) {
 /// Convert any dynamic value to Map<String, dynamic> or null safely.
 Map<String, dynamic>? toMapOrNull(value) {
   if (value == null) return null;
+  if (value is Map && value.isEmpty) return <String, dynamic>{};
   if (value is Map<String, dynamic>) return value;
+  return null;
+}
+
+/// Convert any dynamic value to Map<String, dynamic> or null safely.
+Map<T, S>? toMapOrNullAny<T, S>(value) {
+  if (value == null) return null;
+  if (value is Map<T, S>) return value;
+  if (value is Map && value.isEmpty) return <T, S>{};
   return null;
 }
 
@@ -60,6 +87,7 @@ List? toListOrNull(value) {
 /// Convert any dynamic value to List<T> or null safely.
 List<T>? toListOrNullT<T>(value) {
   if (value == null) return null;
+  if (value is List && value.isEmpty) return [];
   if (value is List<T>) return value;
   return null;
 }
@@ -68,14 +96,31 @@ List<T>? toListOrNullT<T>(value) {
 /// As it converts each item to T using [fromJson] function.
 List<T>? toListOrNullFromJson<T>(value, T Function(dynamic json) fromJson) {
   if (value == null) return null;
-  if (value is List) return value.map((e) => fromJson(e)).toList();
+  if (value is List) {
+    if (value.isEmpty) return [];
+    try {
+      return value.map((e) => fromJson(e)).toList();
+    } catch (e, s) {
+      if (kDebugMode) {
+        log('ERROR IN toListOrNullFromJson', error: e, stackTrace: s);
+      }
+      return null;
+    }
+  }
   return null;
 }
 
 T? toTOrNull<T>(value, {required T Function(dynamic json) fromJson}) {
   if (value == null) return null;
   if (value is T) return value;
-  return fromJson(value);
+  try {
+    return fromJson(value);
+  } catch (e, s) {
+    if (kDebugMode) {
+      log('ERROR IN toTOrNull', error: e, stackTrace: s);
+    }
+    return null;
+  }
 }
 
 /// Convert any dynamic value to int or [fallback] safely.
@@ -126,7 +171,8 @@ List<T> toListFromJsonOr<T>(value, T Function(dynamic json) fromJson,
 }
 
 /// Convert any dynamic value to T or [fallback] safely.
-T toTOr<T>(value, T Function(dynamic json) fromJson, {required T fallback}) {
+T toTOr<T>(value,
+    {required T Function(dynamic json) fromJson, required T fallback}) {
   return toTOrNull(value, fromJson: fromJson) ?? fallback;
 }
 
@@ -190,410 +236,4 @@ List<T> toListFromJson<T>(value, T Function(dynamic json) fromJson) {
 T toT<T>(value, {required T Function(dynamic json) fromJson}) {
   return toTOrNull(value, fromJson: fromJson) ??
       (throw FormatException('Invalid $T value :$value'));
-}
-
-/// Convert any map value with [key] to int or null safely.
-int? asIntOrNull(dynamic json, String key) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-  return toIntOrNull(json[key]);
-}
-
-/// Convert any map value with [key] to double or null safely.
-double? asDoubleOrNull(dynamic json, String key) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-  return toDoubleOrNull(json[key]);
-}
-
-/// Convert any map value with [key] to bool or null safely.
-bool? asBoolOrNull(
-  dynamic json,
-  String key,
-) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-  return toBoolOrNull(json[key]);
-}
-
-/// Convert any map value with [key] to String or null safely.
-String? asStringOrNull(
-  dynamic json,
-  String key,
-) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-  return toStringOrNull(json[key]);
-}
-
-/// Convert any map value with [key] to Map<String, dynamic> or null safely.
-Map<String, dynamic>? asMapOrNull(
-  dynamic json,
-  String key,
-) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-  return toMapOrNull(json[key]);
-}
-
-/// Convert any map value with [key] to List or null safely.
-List? asListOrNull(dynamic json, String key) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-  return toListOrNull(json[key]);
-}
-
-/// Convert any map value with [key] to List<int> or null safely.
-List<int>? asListIntOrNull(
-  dynamic json,
-  String key,
-) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-
-  return asListOrNull(
-    json,
-    key,
-  )?.map((e) => toIntOrNull(e)).whereType<int>().toList();
-}
-
-/// Convert any map value with [key] to List<T> or null safely.
-List<String>? asListStringOrNull(
-  dynamic json,
-  String key,
-) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-  return asListOrNull(
-    json,
-    key,
-  )?.map((e) => toStringOrNull(e)).whereType<String>().toList();
-}
-
-/// Convert any map value with [key] to List<T> or null safely.
-List<T>? asListOrNullT<T>(dynamic json, String key,
-    {required T Function(dynamic json)? fromJson}) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-
-  return asListOrNull(
-    json,
-    key,
-  )?.map((e) => fromJson?.call(e)).whereType<T>().toList();
-}
-
-/// Convert any map value with [key] to T or null safely.
-T? asTOrNull<T>(
-  dynamic json,
-  String key, {
-  T Function(dynamic json)? fromJson,
-}) {
-  if (json == null || json is! Map<String, dynamic> || !json.containsKey(key)) {
-    return null;
-  }
-  dynamic value = json[key];
-
-  if (fromJson != null) {
-    return fromJson(value);
-  }
-
-  if (value is T) return value;
-
-  if (0 is T) {
-    return toIntOrNull(value) as T;
-  } else if (0.0 is T) {
-    return toDoubleOrNull(value) as T;
-  } else if ('' is T) {
-    return toStringOrNull(
-      value,
-    ) as T;
-  } else if (false is T) {
-    return toBoolOrNull(
-      value,
-    ) as T;
-  } else if ([] is T) {
-    return toListOrNull(
-      value,
-    ) as T;
-  } else if (<String, dynamic>{} is T) {
-    return toMapOrNull(
-      value,
-    ) as T;
-  }
-  return null;
-}
-
-/// Convert any map value with [key] to int or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-int asInt(
-  dynamic json,
-  String key,
-) {
-  return toInt(
-    asTOrNull(
-      json,
-      key,
-    ),
-  );
-}
-
-/// Convert any map value with [key] to double or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-double asDouble(
-  dynamic json,
-  String key,
-) {
-  return toDouble(
-    asTOrNull(
-      json,
-      key,
-    ),
-  );
-}
-
-/// Convert any map value with [key] to bool or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-bool asBool(
-  dynamic json,
-  String key,
-) {
-  return toBool(
-    asTOrNull(
-      json,
-      key,
-    ),
-  );
-}
-
-/// Convert any map value with [key] to String or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-String asString(
-  dynamic json,
-  String key,
-) {
-  return toString(
-    asTOrNull(
-      json,
-      key,
-    ),
-  );
-}
-
-/// Convert any map value with [key] to Map<String, dynamic> or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-Map<String, dynamic> asMap(
-  dynamic json,
-  String key,
-) {
-  return toMap(
-    asTOrNull(
-      json,
-      key,
-    ),
-  );
-}
-
-/// Convert any map value with [key] to List or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-List asList(
-  dynamic json,
-  String key,
-) {
-  return toList(
-    asTOrNull(
-      json,
-      key,
-    ),
-  );
-}
-
-/// Convert any map value with [key] to List<T> or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-List<T> asListT<T>(
-  dynamic json,
-  String key,
-) {
-  return toListT(
-    asTOrNull(
-      json,
-      key,
-    ),
-  );
-}
-
-/// Convert any map value with [key] to List<T> or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-List<T> asListFromJson<T>(
-    dynamic json, String key, T Function(dynamic json) fromJson,
-    {List<T> fallback = const []}) {
-  return toListFromJson(
-    asTOrNull(
-      json,
-      key,
-    ),
-    fromJson,
-  );
-}
-
-/// Convert any map value with [key] to List<int> or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-List<int> asListInt(
-  dynamic json,
-  String key,
-) {
-  return asListT(
-    json,
-    key,
-  );
-}
-
-/// Convert any map value with [key] to List<String> or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-List<String> asListString(
-  dynamic json,
-  String key,
-) {
-  return asListT(
-    json,
-    key,
-  );
-}
-
-/// Convert any map value with [key] to T or [fallback] safely.
-/// Throws a [FormatException] if the conversion fails.
-T asT<T>(
-  dynamic json,
-  T Function(dynamic json)? fromJson,
-  String key,
-) {
-  return asTOrNull(json, key, fromJson: fromJson) ??
-      (throw FormatException('Invalid $T value :$key'));
-}
-
-/// Convert any map value with [key] to int or [fallback] safely.
-/// If [fallback] is not provided, it defaults to 0.
-int asIntOr(dynamic json, String key, {int fallback = 0}) {
-  return toIntOr(
-    asTOrNull(
-      json,
-      key,
-    ),
-    fallback: fallback,
-  );
-}
-
-/// Convert any map value with [key] to double or [fallback] safely.
-/// If [fallback] is not provided, it defaults to 0.0.
-double asDoubleOr(dynamic json, String key, {double fallback = 0.0}) {
-  return toDoubleOr(
-      asTOrNull(
-        json,
-        key,
-      ),
-      fallback: fallback);
-}
-
-/// Convert any map value with [key] to bool or [fallback] safely.
-/// If [fallback] is not provided, it defaults to false.
-bool asBoolOr(dynamic json, String key, {bool fallback = false}) {
-  return toBoolOr(
-      asTOrNull(
-        json,
-        key,
-      ),
-      fallback: fallback);
-}
-
-/// Convert any map value with [key] to String or [fallback] safely.
-/// If [fallback] is not provided, it defaults to an empty string.
-String asStringOr(dynamic json, String key, {String fallback = ''}) {
-  return toStringOr(
-      asTOrNull(
-        json,
-        key,
-      ),
-      fallback: fallback);
-}
-
-/// Convert any map value with [key] to Map<String, dynamic> or [fallback] safely.
-/// If [fallback] is not provided, it defaults to an empty map.
-Map<String, dynamic> asMapOr(dynamic json, String key,
-    {Map<String, dynamic> fallback = const <String, dynamic>{}}) {
-  return toMapOr(
-      asTOrNull(
-        json,
-        key,
-      ),
-      fallback: fallback);
-}
-
-/// Convert any map value with [key] to List or [fallback] safely.
-/// If [fallback] is not provided, it defaults to an empty list.
-List asListOr(dynamic json, String key, {List fallback = const []}) {
-  return toListOr(
-      asTOrNull(
-        json,
-        key,
-      ),
-      fallback: fallback);
-}
-
-/// Convert any map value with [key] to List<T> or [fallback] safely.
-/// If [fallback] is not provided, it defaults to an empty list.
-List<T> asListTOr<T>(dynamic json, String key, {List<T> fallback = const []}) {
-  return toListTOr(
-      asTOrNull(
-        json,
-        key,
-      ),
-      fallback: fallback);
-}
-
-/// Convert any map value with [key] to List<T> or [fallback] safely.
-/// If [fallback] is not provided, it defaults to an empty list.
-List<T> asListFromJsonOr<T>(
-    dynamic json, String key, T Function(dynamic json) fromJson,
-    {List<T> fallback = const []}) {
-  return toListFromJsonOr(
-      asTOrNull(
-        json,
-        key,
-      ),
-      fromJson,
-      fallback: fallback);
-}
-
-/// Convert any map value with [key] to List<int> or [fallback] safely.
-/// If [fallback] is not provided, it defaults to an empty list.
-List<int> asListIntOr(dynamic json, String key,
-    {List<int> fallback = const []}) {
-  return asListTOr(json, key, fallback: fallback);
-}
-
-/// Convert any map value with [key] to List<String> or [fallback] safely.
-/// If [fallback] is not provided, it defaults to an empty list.
-List<String> asListStringOr(dynamic json, String key,
-    {List<String> fallback = const []}) {
-  return asListTOr(json, key, fallback: fallback);
-}
-
-/// Convert any map value with [key] to T or [fallback] safely.
-/// If [fallback] is not provided, it defaults to 0.
-T asTOr<T>(dynamic json, String key, T Function(dynamic json)? fromJson,
-    {required T fallback}) {
-  return asTOrNull(
-        json,
-        key,
-        fromJson: fromJson,
-      ) ??
-      fallback;
 }
